# Remove specialized PartialQuickSort{<:Integer} method