@nalimilan 2020-10-02
@kmsquire Do you have any idea why there used to be a specialize method which was actually slower? Was there a mistake in the algorithm?